requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
eventlet==0.33.3
pandas==2.3.3
numpy==2.3.4
openpyxl==3.1.2
Pillow==12.0.0
qrcode==7.4.2
pywin32==311
python-dotenv==1.0.0
dnspython==2.4.2
opencv-python==4.9.0.80